Fix import moves to dbquery file
